src/Deployment/requirements.txt
# AI Road Inspection System Dependencies
# Install with: pip install -r requirements.txt

# Core ML and Computer Vision
ultralytics==8.0.112
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1

# Data Processing
numpy>=1.23.5
pandas>=1.5.0

# Web Framework
streamlit==1.31.0
streamlit_option_menu==0.3.2

# Visualization - MISSING FROM YOUR INSTALL
plotly>=5.0.0

# Image/Video Processing
Pillow==10.4.0
imageio==2.26.1
moviepy==1.0.3

# File Handling
openpyxl>=3.0.0

# Geocoding
geopy>=2.0.0

# HTTP Requests
requests>=2.28.0

# Other utilities
altair>=4.0.0
protobuf==3.20.1